Move sample-app into samples/mvvm.
