--- sample-app/src/main/java/com/lyft/domic/sampleapp/signin/SignInService.kt
@@ -1,15 +1,0 @@
-package com.lyft.domic.sampleapp.signin
-
-import io.reactivex.Observable
-
-interface SignInService {
-
-    data class Credentials(val email: String, val password: String)
-
-    sealed class SignInResult {
-        object Success : SignInResult()
-        data class Error(val cause: Throwable) : SignInResult()
-    }
-
-    fun signIn(credentials: Credentials): Observable<SignInResult>
-}
