+++ samples/mvvm/proguard-rules.pro
@@ -1,0 +1,3 @@
+# Add project specific ProGuard rules here.
+# For more details, see
+# http://developer.android.com/guide/developing/tools/proguard.html
