+++ samples/shared/src/main/java/com/lyft/domic/samples/shared/signin/SignInService.kt
@@ -1,0 +1,15 @@
+package com.lyft.domic.samples.shared.signin
+
+import io.reactivex.Observable
+
+interface SignInService {
+
+    data class Credentials(val email: String, val password: String)
+
+    sealed class SignInResult {
+        object Success : SignInResult()
+        data class Error(val cause: Throwable) : SignInResult()
+    }
+
+    fun signIn(credentials: Credentials): Observable<SignInResult>
+}
