+++ samples/shared/src/main/res/layout/activity_main.xml
