+++ samples/shared/src/main/res/mipmap-mdpi/ic_launcher.png
