+++ samples/shared/src/main/res/mipmap-xxxhdpi/ic_launcher.png
