+++ samples/shared/src/main/res/values/colors.xml
