+++ samples/shared/src/main/res/values/styles.xml
